Modify : Util 함수 추가 및 타입 위치 변경
